+++ CHANGELOG.md
@@ -77,3 +77,4 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 - EmailSenderIfError annotation
 - Frontend DTO converters
 - Fetch all user visible notifications
+- Accepted state requirement for setting user as developer or pilot